style: normalize Moongate foundation formatting
- Normalize C# formatting, member layout, and unused imports across the foundation projects
- Align project file XML spacing after the Moongate import
- Refresh test formatting while preserving the passing plugin config binding behavior

+++ tests/Moongate.Tests/UO/Data/Support/UopFixture.cs
@@ -12,36 +12,37 @@ public static class UopFixture
         string directory,
         string baseName,
         byte[] payload,
-        int version = 5)
+        int version = 5
+    )
     {
         var path = Path.Combine(directory, baseName + ".uop");
 
         using var fs = new FileStream(path, FileMode.Create, FileAccess.Write, FileShare.None);
         using var bw = new BinaryWriter(fs);
 
-        bw.Write((uint)0x50594D);   // "MYP\0"
-        bw.Write(version);          // version
-        bw.Write((uint)0xFD23EC43); // signature
+        bw.Write((uint)0x50594D); // "MYP\0"
+        bw.Write(version);        // version
+        bw.Write(0xFD23EC43);     // signature
 
         var nextBlockPos = fs.Position;
-        bw.Write((long)0);          // placeholder for first block offset
+        bw.Write((long)0); // placeholder for first block offset
 
         var dataOffset = fs.Position;
         bw.Write(payload);
 
         var blockOffset = fs.Position;
-        bw.Write(1);                // fileCount
-        bw.Write((long)0);          // nextBlock = 0 (stop)
+        bw.Write(1);       // fileCount
+        bw.Write((long)0); // nextBlock = 0 (stop)
 
         var hash = UopIndexReader.HashLittle2($"build/{baseName.ToLowerInvariant()}/00000000.dat");
 
-        bw.Write(dataOffset);       // offset
-        bw.Write(0);                // headerLength
-        bw.Write(payload.Length);   // compressedLength
-        bw.Write(payload.Length);   // decompressedLength
-        bw.Write(hash);             // fileNameHash
-        bw.Write((uint)0);          // Adler32
-        bw.Write((short)0);         // compressed flag
+        bw.Write(dataOffset);     // offset
+        bw.Write(0);              // headerLength
+        bw.Write(payload.Length); // compressedLength
+        bw.Write(payload.Length); // decompressedLength
+        bw.Write(hash);           // fileNameHash
+        bw.Write((uint)0);        // Adler32
+        bw.Write((short)0);       // compressed flag
 
         bw.Flush();
         fs.Seek(nextBlockPos, SeekOrigin.Begin);
